+++ R/plot.monthglm.R
@@ -37,11 +37,11 @@ plot.monthglm <- function(
   }
   op <- par(no.readonly = TRUE) # the whole list of settable par's.
   ## y-axis limits
-  if (is.null(ylim) == FALSE) {
-    this.y.lim = ylim
+  if (!is.null(ylim)) {
+    this.y.lim <- ylim
   }
   ## Create CIs
-  refer = NA # reference level - only used for binomial or poisson
+  refer <- NA # reference level - only used for binomial or poisson
   z <- qnorm(1 - (alpha / 2))
   s <- summary(x$glm)
   type <- as.character(x$call$family)[1]
@@ -57,15 +57,15 @@ plot.monthglm <- function(
     out$mean <- exp(out$mean)
     out$lower <- exp(out$lower)
     out$upper <- exp(out$upper)
-    refer = 1
+    refer <- 1
   }
   index <- grep("months", row.names(out), ignore.case = TRUE, value = FALSE)
   toplot <- out[index, ] # Select months
 
   ## Get month names
-  compress <- gsub('months', '', row.names(toplot))
+  compress <- gsub('months', '', row.names(toplot), fixed = TRUE)
   order <- vector(length = nrow(toplot), mode = 'numeric')
-  for (i in 1:nrow(toplot)) {
+  for (i in seq_len(nrow(toplot))) {
     order[i] <- sum(as.numeric(month.abb == compress[i]) * (1:12))
   }
 
@@ -73,7 +73,7 @@ plot.monthglm <- function(
   ymin <- min(c(toplot$lower, refer), na.rm = T) # include reference
   ymax <- max(c(toplot$upper, refer), na.rm = T) # include reference
   if (is.null(ylim)) {
-    this.y.lim = c(ymin, ymax)
+    this.y.lim <- c(ymin, ymax)
   }
   plot(
     order,
@@ -86,7 +86,7 @@ plot.monthglm <- function(
     ...
   )
   month.lab <- vector(mode = 'character', length = nrow(toplot))
-  for (i in 1:nrow(toplot)) {
+  for (i in seq_len(nrow(toplot))) {
     lines(c(order[i], order[i]), c(toplot$lower[i], toplot$upper[i]))
     months.num <- keep_month_numbers(row.names(toplot)[i]) # replaced `nochars`
     month.lab[i] <- month.abb[months.num]