+++ NEWS.md
@@ -8,4 +8,5 @@
 
 ## Breaking change
 
+* `monthmean()` argument, `adjmonth` must be one of "none" (default), "thirty", or "average", rather than: FALSE, "thirty", or "average".
 * `createAdj()` argument, `filepath` has been removed, as it is no longer required.